refactor: article, like 디렉토리 세분화
