+++ backend/src/main/java/com/personal/kopmorning/domain/article/article/responseCode/ArticleSuccessCode.java
@@ -1,4 +1,4 @@
-package com.personal.kopmorning.domain.article.responseCode;
+package com.personal.kopmorning.domain.article.article.responseCode;
 
 import lombok.Getter;
